requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
Pillow==4.2.1
gunicorn==19.7.1
whitenoise==3.3.0
djangorestframework
coverage
